Task 6

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long long int d;
 	float g;
 
-printf("Size of a char: %lu byte(s)", (unsigned long) sizeof(a));
-printf("Size of an int: %lu byte(s)", (unsigned long) sizeof(b));
-printf("Size of a long int: %lu byte(s)", (unsigned long) sizeof(c));
-printf("Size of a long long int: %lu byte(s)", (unsigned long) sizeof(d));
-printf("Size of a float: %lu byte(s)", (unsigned long) sizeof(g));
-return (0);
+	printf("Size of a char: %lu byte(s)", sizeof(a));
+	printf("Size of an int: %lu byte(s)", sizeof(b));
+	printf("Size of a long int: %lu byte(s)", sizeof(c));
+	printf("Size of a long long int: %lu byte(s)", sizeof(d));
+	printf("Size of a float: %lu byte(s)" sizeof(g));
+	return (0);
 }
